==> notebooks/1_prepare_xgb_data_one_signal.ipynb <==
import json
from pathlib import Path

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import anatools.data as data
import anatools.analysis as ana
from sklearn.model_selection import train_test_split

from hhdm_analysis.xgb.controllers import XGBLearner, XGBModel

ana.start()
plt.style.use("default")

# Setup config

period = '18'
year_style = 2018
dataset_year = "2018"
basedir = '/home/gamoreir/SanDisk/physics/hhdmAnalysis/datasets'
dataset_name = basedir.split('/')[-2]

# Setup output folders
data_path = f"./data/{dataset_name}/{dataset_year}"
Path(data_path).mkdir(parents=True, exist_ok=True)

# Read metadata

with open("../metadata.json", "r") as f:
    metadata = json.load(f)

ST = metadata.get("datasets").get("ST")
TT = metadata.get("datasets").get("TT")
ZZ = metadata.get("datasets").get("ZZ")
WZ = metadata.get("datasets").get("WZ")
DY = metadata.get("datasets").get("DY")
RESIDUAL = metadata.get("datasets").get("RESIDUAL")
DATA = metadata.get("datasets").get("DATA")

# Read data

variables = ["RegionID", "evtWeight", "LeadingLep_pt", "LepLep_pt", "LepLep_deltaR", "LepLep_deltaM", "MET_pt", "MET_LepLep_Mt", "MET_LepLep_deltaPhi", "TrailingLep_pt", "MT2LL", "Nbjets"]
ds = data.read_files(basedir, period, mode="normal", features=variables)

data.join_datasets(ds, "ST", ST.get(period), mode="normal")
data.join_datasets(ds, "TT", TT.get(period), mode="normal")
data.join_datasets(ds, "ZZ", ZZ.get(period), mode="normal")
data.join_datasets(ds, "WZ", WZ.get(period), mode="normal")
data.join_datasets(ds, "DYJetsToLL", DY.get(period), mode="normal")
data.join_datasets(ds, "Residual", RESIDUAL.get(period), mode="normal")
data.join_datasets(ds, "Data", DATA.get(period), mode="normal")

print("Signal_400_100", ds["Signal_400_100"].shape)
print("Signal_1000_100", ds["Signal_1000_100"].shape)
print("ST", ds["ST"].shape)
print("TT", ds["TT"].shape)
print("ZZ", ds["ZZ"].shape)
print("WZ", ds["WZ"].shape)
print("DYJetsToLL", ds["DYJetsToLL"].shape)
print("Residual", ds["Residual"].shape)
print("Data", ds["Data"].shape)

del ds["Data"]

# Pre-process datasets

* Filter Signal Region
* XGBoost do not handle negative instance weights (so we remove then)
* Label Background as 0 and Signal as 1
* Weights are normalized (after splitting into train and test) in order to give the same importance for Signal and Background while minimizing the loss function.

# Used for scaling the weights (this helps xgboost)
scale_factor = 10**6

# Filtering SR
ds = {k: v[v.RegionID == 0] for k,v in ds.items()}

# Selecting backgrounds and signals
model_name = "XGB_Signal_400_100"
signal_name = "Signal_400_100"
signal = ds[signal_name]
backgrounds = [ds["ST"], ds["TT"], ds["ZZ"], ds["WZ"], ds["DYJetsToLL"], ds["Residual"]]

# Remove negative values
backgrounds = [bkg[bkg.evtWeight >= 0] for bkg in backgrounds]
signal = signal[signal.evtWeight >= 0]

# Labeling
signal.loc[:, "Label"] = 1
for bkg in backgrounds:
    bkg.loc[:, "Label"] = 0

# Spliting
backgrounds = [train_test_split(bkg, test_size=0.3, random_state=42) for bkg in backgrounds]
X_train_signal, X_test_signal = train_test_split(signal, test_size=0.3, random_state=42)

# Group backgrounds
X_train_bkgs = pd.concat([X_train_bkg for X_train_bkg, _ in backgrounds])
X_test_bkgs = pd.concat([X_test_bkg for _, X_test_bkg in backgrounds])

# Normalize background weights
X_train_bkgs["modelWeight"] = scale_factor*(X_train_bkgs.evtWeight/X_train_bkgs.evtWeight.sum())
X_test_bkgs["modelWeight"] = scale_factor*(X_test_bkgs.evtWeight/X_test_bkgs.evtWeight.sum())

# Normalize signal weights
X_train_signal["modelWeight"] = scale_factor*(X_train_signal.evtWeight/X_train_signal.evtWeight.sum())
X_test_signal["modelWeight"] = scale_factor*(X_test_signal.evtWeight/X_test_signal.evtWeight.sum())

# Group signal and background

# Group signal and background
X_train = pd.concat([X_train_signal, X_train_bkgs]).reset_index(drop=True)
X_test = pd.concat([X_test_signal, X_test_bkgs]).reset_index(drop=True)

# Normalize weights
X_train["modelWeight"] = scale_factor*(X_train.modelWeight/X_train.modelWeight.sum())
X_test["modelWeight"] = scale_factor*(X_test.modelWeight/X_test.modelWeight.sum())

# Shuffle
X_train = X_train.sample(frac=1).reset_index(drop=True)
X_test = X_test.sample(frac=1).reset_index(drop=True)

print(X_train.shape)
X_train.head()

print(X_test.shape)
X_test.head()

# Save processed data

# Pop unused variables at this point
X_train.pop("RegionID")
X_test.pop("RegionID")
# X_train.pop("evtWeight")
# X_test.pop("evtWeight")

X_train.to_csv(f"{data_path}/{model_name}-train-data.csv", index=False)
X_test.to_csv(f"{data_path}/{model_name}-test-data.csv", index=False)
